Video #17 Validation/Requirements when entering data into forms fixed with help of Dto:s (look there) and into the controllers by just adding some small [:int] and ModelState parameter, works nice

+++ api/Controllers/CommentController.cs
@@ -11,7 +11,7 @@ namespace api.Controllers
 {
     [Route("api/comment")]          // Defines the route for the controller
     [ApiController]                 // Defines the controller as an API controller
-    public class CommentController : ControllerBase
+    public class CommentController : ControllerBase     // Foreign Key controller, Comments
     {
         private readonly ICommentRepository _commentRepo;
         private readonly IStockRepository _stockRepo;
@@ -23,9 +23,12 @@ namespace api.Controllers
         }
 
 //-----------------
-        [HttpGet]                   // GetAll
+        [HttpGet]                       // GetAll
         public async Task<IActionResult> GetAll()
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -33,9 +36,12 @@ namespace api.Controllers
             return Ok(commentDto);
         }
 
-        [HttpGet("{id}")]           // GetById
+        [HttpGet("{id:int}")]           // GetById
         public async Task<IActionResult> GetById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var comment = await _commentRepo.GetByIdAsync(id);
 
             if (comment == null)
@@ -46,9 +52,12 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("{stockId}")]     // Create
+        [HttpPost("{stockId:int}")]     // Create
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             if(!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist");
@@ -59,10 +68,13 @@ namespace api.Controllers
             return CreatedAtAction(nameof(GetById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
-        [HttpPut]                   // Update
-        [Route("{id}")]
+        [HttpPut]                       // Update
+        [Route("{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate());
 
             if (comment == null)
@@ -73,10 +85,13 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpDelete]                // Delete
-        [Route("{id}")]
+        [HttpDelete]                    // Delete
+        [Route("{id:int}")]
         public async Task<IActionResult> Delete([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+                
             var commentModel = await _commentRepo.DeleteAsync(id);
 
             if (commentModel == null)
